example_3/bots/requirements.txt
pyqalx==0.18.6
python-docx==0.8.11